requirements.txt
qdrant-client
openai
sentence-transformers
pandas
numpy
python-dotenv
requests